+++ timeseriestable/src/TimeSeriesTablePanel.tsx
@@ -17,7 +17,7 @@ import { ReactElement } from 'react';
 import { PanelProps } from '@perses-dev/plugin-system';
 import { TimeSeriesData } from '@perses-dev/core';
 import { TimeSeriesTableOptions } from './model';
-import { DataTable } from './DataTable';
+import { DataTable } from './components';
 
 export type TimeSeriesTableProps = PanelProps<TimeSeriesTableOptions, TimeSeriesData>;
