+++ timeseriestable/src/components/SeriesName.tsx
