+++ timeseriestable/src/components/index.ts
@@ -1,1 +1,3 @@
+export * from './DataTable';
 export * from './HistogramChart';
+export * from './SeriesName';
